[refactor] Use camelCase for aptosConfig
