+++ src/api/index.ts
@@ -2,4 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from "./aptos";
-export * from "./aptos_config";
+export * from "./aptosConfig";
