+++ src/api/token.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosConfig } from "./aptos_config";
+import { AptosConfig } from "./aptosConfig";
 import { getTokenData } from "../internal/token";
 import { GetTokenDataResponse, HexInput } from "../types";
