+++ src/client/post.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosConfig } from "../api/aptos_config";
+import { AptosConfig } from "../api/aptosConfig";
 import { aptosRequest } from "./core";
 import { AptosResponse } from "./types";
 import { AnyNumber, ClientConfig, MimeType } from "../types";
