+++ src/internal/account.ts
@@ -8,7 +8,7 @@
  * account namespace and without having a dependency cycle error.
  */
 
-import { AptosConfig } from "../api/aptos_config";
+import { AptosConfig } from "../api/aptosConfig";
 import { AptosApiError, getAptosFullNode, paginateWithCursor } from "../client";
 import { AccountAddress, Hex } from "../core";
 import { getTableItem, queryIndexer } from "./general";
