+++ src/internal/faucet.ts
@@ -8,7 +8,7 @@
  * faucet namespace and without having a dependency cycle error.
  */
 
-import { AptosConfig } from "../api/aptos_config";
+import { AptosConfig } from "../api/aptosConfig";
 import { postAptosFaucet } from "../client";
 import { AccountAddress } from "../core";
 import { HexInput } from "../types";
