+++ src/internal/general.ts
@@ -8,7 +8,7 @@
  * general namespace and without having a dependency cycle error.
  */
 
-import { AptosConfig } from "../api/aptos_config";
+import { AptosConfig } from "../api/aptosConfig";
 import { getAptosFullNode, postAptosFullNode, postAptosIndexer } from "../client";
 import {
   AnyNumber,
