+++ src/internal/transaction.ts
@@ -8,7 +8,7 @@
  * transaction namespace and without having a dependency cycle error.
  */
 
-import { AptosConfig } from "../api/aptos_config";
+import { AptosConfig } from "../api/aptosConfig";
 import { AptosApiError, getAptosFullNode, paginateWithCursor } from "../client";
 import {
   TransactionResponseType,
